# apps/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AIO Creative Hub API
  description: |
    API for the AIO Creative Hub - a chat-powered creative platform for design, coding, 3D modeling, and video editing.

    ## Features
    - Natural language processing for creative tool routing
    - Real-time collaboration via WebSocket
    - Multi-tool creative workflow support
    - Google Drive integration
    - Secure authentication with OAuth 2.0

    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Standard endpoints: 100 requests per minute
    - Tool operations: 60 requests per minute
    - File uploads: 20 requests per minute

    ## WebSocket
    The API supports real-time communication via WebSocket. Connect to:
    ```
    ws://localhost:3000/ws
    ```

    ## Error Codes
    - 400: Bad Request - Invalid input
    - 401: Unauthorized - Authentication required
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource not found
    - 429: Too Many Requests - Rate limit exceeded
    - 500: Internal Server Error

  version: 1.0.0
  contact:
    name: AIO Creative Hub API Support
    email: support@aio-creative-hub.com
    url: https://aio-creative-hub.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "securePassword123"
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refreshToken:
                    type: string
        '400':
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refreshToken:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/google:
    get:
      tags:
        - Authentication
      summary: Google OAuth login
      description: Initiate Google OAuth 2.0 authentication flow
      operationId: googleAuth
      responses:
        '302':
          description: Redirect to Google OAuth
        '400':
          description: Invalid request
  /api/auth/google/callback:
    get:
      tags:
        - Authentication
      summary: Google OAuth callback
      description: Handle Google OAuth callback
      operationId: googleCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '400':
          description: Authentication failed
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Get a new JWT token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                  refreshToken:
                    type: string
        '401':
          description: Invalid refresh token
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current session
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true

  # User Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...